+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 
 * New `:id` option allows you to identify the client against Redis. See #510.
 
+* `Redis::Distributed` will raise when adding two nodes with the same ID.
+  See #354.
+
 # 3.2.1
 
 * Added support for `PUBSUB` command.